introduced constructors, sketched Board class
I made instance fields private and make getters and setters for them where appropriate. i made the pieces classes call the constructor in Piece. I also sketched out the board class and implemented a method to check if a given coordinate is in the bounds of the board. the Chess class now has a static variable ChessBoard. I also simplified the move input formatting a bit.