moved common canMove() logic to Piece class
now we can call canMove() on any arbitrary Piece with given coordinates and it should return whether or not it can move to that location. this method should also handle the logic for checking if there are pieces in the way, if the new square is valid, if the piece is the right color, etc.