Castling works (in the testing I did)
Castling required a little bit of extra logic for the king's position. additional variables included to handle accordingly. Some code removed and changed due to isSquareVisibleByEnemy limitations and replaced with a canMove() call in a loop for all enemy pieces. Tested and pieces seem to castle correctly, but I might need to retest this.