inCheck() works....?
Current implementation of inCheck() works for the few cases I threw at it. Currently, a major issue with inCheck() and potentially promotion is that PlayChess.java cannot handle the null elements that may appear in the piecesOnBoard arraylist (I believe this is due to how resizing arrayLists works). Make sure to add a line of code to remove nulls before returns if you get any errors from it. But inCheck() seems to work.