fixed an issue with movePiece()
movePiece would move the piece back if the piece caused the king to be checked. the problem was this was interfering with isInCheckmate, which uses movePiece() to temporarily move pieces to check all possible squares. so i moved the check that moves the piece back into Chess.play, which should fix the issue.