READ MESSAGE: Castling Bug Fixes
MAJOR ISSUE WITH CASTLING FIXED. Previously, moves would specify king to rook's position for castling. This is INCORRECT CASTLING FORMATTING (according to course site). Move must be indicated by specifying where king begins and ends (e1 g1, not e1 h1). This is NOT GOOD.

Dumb solution: offset. offset the castle whenever castling input is given (and we check that we're moving a king first)

This is the second version of this, so I HOPE it doesn't break anything. This time.